Bug #647532: Enter APR as interest rate in Loan Druid
Patch by Nick:

This patch allows the user to enter the APR for a loan rather than the simple
interest rate and then the druid will calculate the correct interest rate for
the scheduled transaction that it produces.

It includes a small change to the shed-xact.glade file which provides a
dropdown box to select the rate that has been entered and changes to
druid-loan.c which calculate the simple interest rate as the scheduled
transaction strings are generated.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@20646 57a11ea4-9604-0410-9ed3-97b8803252fd