Bug #650757: Fix Easy & footer styles output wrong tags </btd> iso </b></td> preventing HTML 4.01 Transitional validation
Patch by Bert:

This patch removes the wrong </btd> tags and closes the center tag in the
footer and easy stylesheets.

Just like Bug #650163 (and related to bug #616611) the easy & footer
stylesheets:
*output </btd> tags instead of <b></td>
*are missing a </center> tag

preventing html validation.

This patch removes the intended </b> and makes a clean </td> out of the
necessary tags. The styles are set through css anyhow. Furthmore it adds the
closing </center> tag.
BP

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@20675 57a11ea4-9604-0410-9ed3-97b8803252fd